+++ CHANGELOG.md
@@ -5,3 +5,8 @@ All notable changes to `laravel-carbon-holidays` will be documented in this file
 ## 1.0.0 - 2021-05-18
 
 - initial release
+
+## 1.0.1 - 2021-05-18
+
+- Goede vrijdag toegevoegd
+- Bevreidingsdag (5 mei) is altijd een feestdag (niet altijd vrij)